refactor: split actions.rs into domain submodules and improve robustness
Break the monolithic 1145-line perform_action match into focused submodules:
- actions/cursor.rs: cursor movement and scroll
- actions/edit.rs: text editing, selection, mode transitions
- actions/buffer.rs: buffer lifecycle (open, save, close, list)
- actions/lsp_actions.rs: LSP requests (hover, completion, definitions, etc.)
- actions/clipboard.rs: register and system clipboard operations
- actions/search.rs: search and find operations
- actions/audio_actions.rs: TTS and transcription

Also:
- Add sync_cursor/extend_selection helpers to BufferInstance, eliminating
  ~25 repeated 4-line cursor/selection sync patterns
- Replace ~30 bare buffer_idx.unwrap() with ? to prevent panics
- Replace .unwrap() on LSP send_request_sync and clipboard operations
  with proper error logging
- Remove 3 empty no-op action variants (InsertAfterSelection,
  SelectTillStartOfWord, ExtendSelectTillStartOfWord)
